requirements.txt
Pillow>=9.0.0
pillow-heif>=1.0.0
